// src/types/contract.ts
export interface TransferContract {
  to_address: string;
  owner_address: string;
  amount: number;
}